fix: Change how algorithm controls are assigned.
When setting algorithm controls in BST, it could not be reused in the AVL-quiz class. This fix allows for algorithm controls to be initialized after the class is initialized. A engine algorithm controls is always defined, and in most cases the engine sub class doesn't need access to the algorithm controls.